third_party/blink/web_tests/accessibility/textarea-selection-expected.txt

This tests that text selection is reported correctly for textarea elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 0
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 25
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 0
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 7
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 7
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 0
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 7
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 14
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 14
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 7
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 14
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 21
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 21
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 14
PASS emptyTextareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 0
PASS emptyTextareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 0
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 21
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 14
PASS emptyTextareaAccessible is emptyTextareaAccessible.selectionAnchorObject
PASS emptyTextareaAccessible.selectionAnchorOffset is 0
PASS emptyTextareaAccessible is emptyTextareaAccessible.selectionFocusObject
PASS emptyTextareaAccessible.selectionFocusOffset is 0
PASS emptyTextareaAccessible.selectionAnchorOffset is 0
PASS emptyTextareaAccessible.selectionFocusOffset is 0
PASS successfullyParsed is true

TEST COMPLETE

 
